Multiple high priority bug fixes
The following issues were fixed:

Excel files where the text is abridged or squished is coming in literally as it would be displayed in Excel.  (i.e. a date inside a column too small appears as #######, and comes through as the literal text of #######)

Intrepid files are sometimes importing without the last column.  This has something to do with the file format, as when they are re-saved in Excel they work fine..  Scott S and Scott D are working together to figure out what can be done to correct.

The SSIS process doesn't handle more than 12-15 files before Excel stops working.

The Process Agency Medical Records segment of the process does not process multiple file format types at once.  If multiple agencies have files in the queue, only the first agency alphabetically will process.